add overflow-x auto

+++ app/controllers/api/v1/problems_controller.rb
@@ -78,10 +78,12 @@ skip_before_action :verify_authenticity_token
     
     def destroy
         problem = Problem.find(params[:id])
-        if problem.destroy
-            render json: {message: '削除しました'}
-        else
-            render json: {error: problem.errors}, status: 422
+        if problem.user.logged_in 
+            if problem.destroy
+                render json: {message: '削除しました'}
+            else
+                render json: {error: problem.errors}, status: 422
+            end
         end
     end
